Implement get_fd_status system call to monitor is fd status
This is used to check if a task requests to read/write a file.
The original idea is to create a task to monitor reading request
to "/proc/stat". If so, the task writes task information back
to "/proc/stat"

Unfortunately, no serial output to console. No idea how to
identify where the problem is. Suggestions are welcomed.